Renamed env to sagemaker.
This avoids it being catched by gitignore.
